pyproject.toml
[project]
name = "zaza-interior-website"
version = "0.1.0"
description = "Website"
requires-python = "==3.12.*"
dependencies = [
    "asgiref==3.8.1",
    "certifi==2025.4.26",
    "chardet==5.2.0",
    "charset-normalizer==3.4.2",
    "cssselect2==0.8.0",
    "diff-match-patch==20241021",
    "django==5.2.1",
    "django-filer==3.3.1",
    "django-import-export==4.3.7",
    "django-polymorphic==3.1.0",
    "django-recaptcha==4.1.0",
    "easy-thumbnails==2.10",
    "et-xmlfile==2.0.0",
    "gunicorn==23.0.0",
    "idna==3.10",
    "lxml==5.4.0",
    "openpyxl==3.1.5",
    "packaging==25.0",
    "pillow==11.2.1",
    "pip==25.1.1",
    "python-decouple==3.8",
    "reportlab==4.4.0",
    "requests==2.32.3",
    "setuptools==80.3.1",
    "sqlparse==0.5.3",
    "svglib==1.5.1",
    "tablib==3.8.0",
    "tinycss2==1.4.0",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "webencodings==0.5.1",
    "wheel==0.45.1",
]
